+++ asfsmd/cli.py
@@ -1,266 +1,25 @@
-#!/usr/bin/env python3
+"""Command Line Interface (CLI) for the ASF S1 Metadata Download tool."""
+
 # PYTHON_ARGCOMPLETE_OK
 
-"""ASF Sentinel-1 Metadata Download tool.
-
-Small Python tool (`asfsmd`) that allows to download XML files containing
-Sentinel-1 products metadata from the ASF archive.
-
-Sentinel-1 products are stored in the ASF arcive as ZIP files that are
-quite large because they comntain both the products annotations and the
-binary image data.
-
-The `asfsmd` tool is able to retrieve only the relatively small annotation
-files (in XML format) without downloading the entire ZIP archive.
-"""
-
 import os
-import sys
 import json
-import netrc
-import fnmatch
 import logging
 import pathlib
-import zipfile
 import argparse
-import warnings
-import functools
-import contextlib
 import collections
 
-from typing import Dict, List, NamedTuple, Optional, Union
-from urllib.parse import urlparse
+from typing import Dict, List, Union
 
 import tqdm
-import asf_search as asf
 
+from . import __version__, __doc__ as DOC
+from .core import (
+    download_annotations, download_components_from_urls, make_patterns,
+    _get_auth,
+)
+from .common import BLOCKSIZE, MB
 
-__version__ = "1.2.0.dev0"
-__all__ = ["download_annotations", "main"]
-
-
-_log = logging.getLogger(__name__)
-
-
-MB = 1024 * 1024
-BLOCKSIZE = 16 * MB  # 16MB (64MB is a better choice to download data)
-
-
-class Auth(NamedTuple):
-    user: str
-    pwd: str
-
-
-PathType = Union[str, bytes, os.PathLike]
-Url = str
-
-
-def query(products):
-    """Query the specified Sentinel-1 products."""
-    if isinstance(products, str):
-        products = [products]
-    products = [product + "-SLC" for product in products]
-    results = asf.product_search(products)
-    return results
-
-
-def make_patterns(
-    beam="*", pol="??", cal=False, noise=False, rfi=False, data=False,
-):
-    """Generate a list of patterns according to the specified options.
-
-    Patterns are used to match components in the ZIP archive of the
-    Sentinel-1 products.
-    """
-    beam = "*" if beam is None else beam
-    pol = "??" if pol is None else pol
-
-    patterns = [
-        "S1*.SAFE/manifest.safe",
-    ]
-
-    head = "S1*.SAFE/annotation"
-    tail = f"s1?-{beam}-???-{pol}-*.xml"
-
-    patterns.append(f"{head}/{tail}")
-    if cal:
-        patterns.append(f"{head}/calibration/calibration-{tail}")
-    if noise:
-        patterns.append(f"{head}/calibration/noise-{tail}")
-    if rfi:
-        patterns.append(f"{head}/rfi/rfi-{tail}")
-
-    if data:
-        patterns.append(f"S1*.SAFE/measurement/s1?-{beam}-???-{pol}-*.tiff")
-
-    return patterns
-
-
-def _filter_components(
-    zf: zipfile.ZipFile, patterns: List[str],
-) -> List[zipfile.ZipInfo]:
-    components = []
-    for info in zf.filelist:
-        for pattern in patterns:
-            if fnmatch.fnmatch(info.filename, pattern):
-                components.append(info)
-                break
-    return components
-
-
-def _download(
-    zf: zipfile.ZipFile, info: zipfile.ZipInfo, outfile: PathType,
-    block_size: int = BLOCKSIZE,
-):
-    size = info.file_size
-    with tqdm.tqdm(total=size, leave=False, unit_scale=True, unit="B") as pbar:
-        with zf.open(info) as src, open(outfile, "wb") as dst:
-            for data in iter(functools.partial(src.read, block_size), b""):
-                dst.write(data)
-                pbar.update(len(data))
-
-
-import io
-
-import httpio
-import requests
-
-
-class HttpIOFile(httpio.SyncHTTPIOFile):
-    def open(self, session=None):
-        self._assert_not_closed()
-        if not self._closing and self._session is None:
-            self._session = requests.Session() if session is None else session
-            response = self._session.get(self.url, stream=True, **self._kwargs)
-            with response:
-                response.raise_for_status()
-                try:
-                    self.length = int(response.headers["Content-Length"])
-                except KeyError:
-                    raise httpio.HTTPIOError(
-                        "Server does not report content length"
-                    )
-                accept_ranges = response.headers.get("Accept-Ranges", "")
-                if accept_ranges.lower() != "bytes":
-                    raise httpio.HTTPIOError(
-                        "Server does not accept 'Range' headers"
-                    )
-        return self
-
-
-class HttpIOClient(contextlib.AbstractContextManager):
-    def __init__(self, auth: Auth, block_size: int = BLOCKSIZE):
-        self._session = requests.Session()
-        self._session.auth = auth
-        self._block_size = block_size
-    
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        self._session.close()
-
-    def open(self, url: Url) -> io.BufferedIOBase:
-        remote_file = HttpIOFile(url, block_size=self._block_size)
-        return remote_file.open(session=self._session)
-
-
-def _get_client(auth: Auth, block_size: Optional[int] = BLOCKSIZE):
-    return HttpIOClient(auth=auth, block_size=block_size)
-
-
-@contextlib.contextmanager
-def _open_zip_archive(url: Url, client) -> zipfile.ZipFile:
-    with client.open(url) as fd:
-        with zipfile.ZipFile(fd) as zf:
-            yield zf
-
-
-def download_components_from_urls(
-    urls, *, patterns: Optional[List[str]] = None, outdir: PathType = ".",
-    auth: Auth = None, block_size: Optional[int] = BLOCKSIZE,
-):
-    """Download Sentinel-1 annotation for the specified product urls."""
-    outdir = pathlib.Path(outdir)
-    if patterns is None:
-        patterns = make_patterns()
-
-    with _get_client(auth=auth, block_size=block_size) as client:
-        url_iter = tqdm.tqdm(urls, unit=" products")
-        for url in url_iter:
-            url_iter.set_description(url)
-            product_name = pathlib.Path(urlparse(url).path).stem
-            _log.debug("download: %r", product_name)
-
-            with _open_zip_archive(url, client) as zf:
-                _log.debug("%s open", url)
-                components = _filter_components(zf, patterns)
-                component_iter = tqdm.tqdm(
-                    components, unit="files", leave=False
-                )
-                for info in component_iter:
-                    filename = pathlib.Path(info.filename)
-                    component_iter.set_description(filename.name)
-                    targetdir = outdir / filename.parent
-                    outfile = targetdir / filename.name
-                    _log.debug("targetdir = %r", targetdir)
-                    _log.debug("outfile = %r", outfile)
-                    targetdir.mkdir(exist_ok=True, parents=True)
-                    if outfile.exists():
-                        _log.debug("outfile = %r exists", outfile)
-                    else:
-                        _download(zf, info, outfile, block_size=block_size)
-                        _log.debug("%r extracted", info.filename)
-
-
-def download_annotations(
-    products: List[str], *, patterns: Optional[List[str] ] = None,
-    outdir: PathType = ".", auth: Auth = None,
-    block_size: Optional[int] = BLOCKSIZE,
-):
-    """Download annotations for the specified Sentinel-1 products."""
-    results = query(products)
-    if len(results) != len(products):
-        warnings.warn(
-            f"only {len(results)} of the {len(products)} requested products "
-            f"found on the remote server"
-        )
-
-    urls = [item.properties["url"] for item in results]
-
-    download_components_from_urls(
-        urls, patterns=patterns, outdir=outdir, auth=auth,
-        block_size=block_size,
-    )
-
-
-def _get_auth(
-    user: Optional[str] = None,
-    pwd: Optional[str] = None,
-    hostname: Url = "urs.earthdata.nasa.gov"
-) -> Auth:
-    if user is not None and pwd is not None:
-        return Auth(user, pwd)
-    elif user is None and pwd is None:
-        db = netrc.netrc()
-        user, _, pwd = db.authenticators(hostname)
-        return Auth(user, pwd)
-    else:
-        raise ValueError(
-            "Both username and password must be provided to authenticate."
-        )
-
-
-# === CLI support functions ===================================================
-def _read_from_file(filename: os.PathLike) -> Union[List[str], Dict[str, str]]:
-    filename = pathlib.Path(filename)
-    if filename.suffix == '.json':
-        return json.loads(filename.read_text())
-    else:
-        return [line for line in filename.read_text().splitlines() if line]
-
-
-# === Command Line Interface ==================================================
 try:
     from os import EX_OK
 except ImportError:
@@ -269,6 +28,14 @@ EX_FAILURE = 1
 EX_INTERRUPT = 130
 
 LOGFMT = "%(asctime)s %(levelname)-8s -- %(message)s"
+
+
+def _read_from_file(filename: os.PathLike) -> Union[List[str], Dict[str, str]]:
+    filename = pathlib.Path(filename)
+    if filename.suffix == '.json':
+        return json.loads(filename.read_text())
+    else:
+        return [line for line in filename.read_text().splitlines() if line]
 
 
 def _autocomplete(parser):
@@ -321,9 +88,9 @@ def _set_logging_control_args(parser, default_loglevel="WARNING"):
 
 def _get_parser(subparsers=None):
     """Instantiate the command line argument (sub-)parser."""
-    name = pathlib.Path(__file__).stem
+    name = __package__
     synopsis = __doc__.splitlines()[0]
-    doc = __doc__
+    doc = DOC
 
     if subparsers is None:
         parser = argparse.ArgumentParser(prog=name, description=doc)
@@ -470,6 +237,7 @@ def main(*argv):
     # setup logging
     logging.basicConfig(format=LOGFMT, level=logging.INFO)  # stream=sys.stdout
     logging.captureWarnings(True)
+    _log = logging.getLogger(__name__)
 
     # parse cmd line arguments
     args = _parse_args(argv if argv else None)
@@ -535,7 +303,3 @@ def main(*argv):
         exit_code = EX_INTERRUPT
 
     return exit_code
-
-
-if __name__ == "__main__":
-    sys.exit(main())